move files.ts to utils folder
